+++ CHANGES.md
@@ -2,6 +2,12 @@
 
 ## 2.6.0
 
+#### Added
+
+- `boot.pod/this-pod` &mdash; a `WeakReference` to the current pod
+- `boot.pod/with-invoke-in` &mdash; low-level invocation, no serialization
+- `boot.pod/with-invoke-worker` &mdash; as above but invokes in the worker pod
+
 ## 2.5.3
 
 #### Improved